Disable norms on fields used to store LSH hashes (#78)
I setup a benchmark for the SparseIndexed mapping/query and noticed a lot of time spent in the Lucene80NormsProducer class. Setting .omitNorms(true) in a few places seems to get rid of that class in the sampler. It decreases the time for the ExecuteLocal benchmark from 100 seconds for 1000 serial queries to ~85 seconds for 1000 serial queries and 74 seconds to 62 seconds for parallel queries.